Exception handling

+++ src/main/java/com/infordevtech/dscommerce/services/ProductService.java
@@ -3,20 +3,32 @@ package com.infordevtech.dscommerce.services;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-
+import com.infordevtech.dscommerce.DscommerceApplication;
 import com.infordevtech.dscommerce.dto.ProductDTO;
 import com.infordevtech.dscommerce.entities.Product;
 import com.infordevtech.dscommerce.repositories.ProductRepository;
+import com.infordevtech.dscommerce.services.exception.DatabaseException;
+import com.infordevtech.dscommerce.services.exception.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
+
+    private final DscommerceApplication dscommerceApplication;
 	
 	@Autowired
 	private ProductRepository repository;
+
+    ProductService(DscommerceApplication dscommerceApplication) {
+        this.dscommerceApplication = dscommerceApplication;
+    }
 	
 	/*
 	@Transactional(readOnly = true)
@@ -34,7 +46,7 @@ public class ProductService {
 	@Transactional(readOnly = true)
 	public ProductDTO findById(Long id) {
 		Optional<Product> result = repository.findById(id);
-		Product product = result.get();
+		Product product = result.orElseThrow(() -> new ResourceNotFoundException("Product not found!"));
 		ProductDTO dto = new ProductDTO(product);
 		return dto;								 
 	}
@@ -63,16 +75,30 @@ public class ProductService {
 	}
 	
 	@Transactional
-	public ProductDTO update(Long id, ProductDTO dto) {		
-		Product entity = repository.getReferenceById(id);
-		copyDtoToEntity(dto, entity);
-		entity = repository.save(entity);
-		return new ProductDTO(entity);
+	public ProductDTO update(Long id, ProductDTO dto) {
+		try {
+			Product entity = repository.getReferenceById(id);
+			copyDtoToEntity(dto, entity);
+			entity = repository.save(entity);
+			return new ProductDTO(entity);
+		}
+		
+		catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Product not found!"); 
+		}
 	}
 	
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		repository.deleteById(id);							 
+		if(!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Product not Found!");
+		}
+		try {
+			repository.deleteById(id);
+		}
+		catch (DataIntegrityViolationException e) {
+			throw new DatabaseException("Referential integrity failure!");
+		}
 	}
 
 	private void copyDtoToEntity(ProductDTO dto, Product entity) {
